\name{RSCC-package}
\alias{RSCC-package}
\alias{RSCC}
\docType{package}
\title{
\packageTitle{RSCC}
}
\description{
\packageDescription{RSCC}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{RSCC}
\packageIndices{RSCC}
~~ An overview of how to use the package, including the most important functions ~~
Main function: 
}
\author{
\packageAuthor{RSCC}

Maintainer: \packageMaintainer{RSCC}
}
\references{
Sparse Spatially Clustered Coefficient Model via Adaptive Regularization}

\keyword{Adaptive regularization, Spatial variable selection, Variable-dependent graph, Varying coefficient models.}

