% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSCC.R
\name{RSCC_gic}
\alias{RSCC_gic}
\title{RSCC with model selection by GIC.}
\usage{
RSCC_gic(
  y,
  X,
  Graph0,
  penalty_factor = NULL,
  intercept = T,
  lambda_row_list = exp(seq(0, -8, length.out = 10)),
  lambda_col_list = exp(seq(0, -8, length.out = 10)),
  lambda_row_list_ini = exp(seq(0,-7,length.out = 10)), 
  lambda_col_list_ini= exp(seq(-3,-8,length.out = 10)) , 
  min_var_ini = dim(X)[2]/2,
  step_size = 1,
  tol = 10^(-7),
  MaxIt = 1000
)
}
\arguments{
\item{y}{The response variable for n observations.}

\item{X}{An n*p matrix for n observations with p covariates. (Do not include intercept)}

\item{Graph0}{A list that contains the information for graph. Graph0$graph is a igraph object that contains the whole graph. Graph0$edge is a three-column matrix includes edge indexes in the first two columns and the distance between nodes of edges in the third column. For unweighted graph, set the third column of Graph0$edge to all 1s. An example of this object is the output of RSCC::buildKNN(). The input graph should be connected.}

\item{penalty_factor}{A vector of length p that indicates the variables that are pre-included into the model. 0 means that the vairable must be included in the model.}

\item{intercept}{Include fixed intercept or not. }

\item{lambda_row_list}{A list of lambda_row for clustering.}

\item{lambda_col_list}{A list of lambda_col for variable selection.}

\item{lambda_row_list_ini,lambda_col_list_ini}{Lists of parameters for model initialization.}

\item{min_var_ini}{An integer requires that the number of variables remained by model initialization should not below this value.}

\item{step_size}{A tuning parameter to adjust step_size. It should be in (0,1].}

\item{tol}{Tolerance for convergence.}

\item{MaxIt}{Max number of iterations.}
}
\value{
Return a list with the estimator for B (Bnow), mse, GIC, the values of three parameters, and the estimated degree of freedom.
}
\description{
RSCC with model selection by gic.
}
