% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same_file.R
\name{same_file}
\alias{same_file}
\title{same_file}
\usage{
same_file(m, replacement = 0)
}
\arguments{
\item{m}{matrix object with row- and columnnames}

\item{replacement}{value for replacement (default: \code{0})}
}
\value{
matrix
}
\description{
same_file
}
\examples{
m <- matrix(runif(25), ncol=5)
colnames(m) <- rownames(m) <- c(sprintf("m[\%.f]", 1:3), sprintf("m2[\%.f]", 1:2))
m
same_file(m)
}
