% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_coeff.R
\name{sim_coeff}
\alias{sim_coeff}
\title{sim_coeff}
\usage{
sim_coeff(set1, set2, setfull, coeff)
}
\arguments{
\item{set1}{character: unique vector of words}

\item{set2}{character: unique vector of words}

\item{setfull}{character: unique vector of texts to compare}

\item{coeff}{character: name of similarity coefficient to use}
}
\value{
value of similarity coefficient
}
\description{
Internal function for faster computation. No checks on input will be performed.
}
