% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcecode.R
\name{sourcecode}
\alias{sourcecode}
\alias{sourcecode.default}
\title{sourcecode}
\usage{
sourcecode(x, ...)

\method{sourcecode}{default}(x, basename = FALSE, silent = FALSE, minlines = -1, ...)
}
\arguments{
\item{x}{character: filenames}

\item{...}{unused}

\item{basename}{logical: should only the basename used for sourcecode objects (default: \code{FALSE})}

\item{silent}{logical: should the report of messages be suppressed (default: \code{FALSE})}

\item{minlines}{integer: only expressions with \code{minlines} lines are considered (default: \code{-1}), if \code{minlines<0} then whole files will be considered}
}
\value{
a sourcecode object
}
\description{
reads and parses files with R source code
}
\examples{
# example files are taken from https://CRAN.R-project.org/package=SimilaR
files <- list.files(system.file("examples", package="rscc"), "*.R$", full.names=TRUE)
prgs  <- sourcecode(files)
}
