% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{browse}
\usage{
browse(prgs, simdf, n = sum(simdf[, 3] > 0), width.cutoff = 60, css = NULL)
}
\arguments{
\item{prgs}{sourcecode object}

\item{simdf}{similarity object}

\item{n}{integer: number of comparisons to show (default: \code{sum(simf[,3]>0)})}

\item{width.cutoff}{integer: an integer in [20, 500]: if a line's character length is at or over this number, the function will try to break it into a new line (default: \code{60})}

\item{css}{character: file name of CSS style for highlighting the R code}
}
\value{
invisibly the name of the temporary HTML file
}
\description{
Creates a temporary HTML file and opens it into a browser.
}
\examples{
# example files are taken from https://CRAN.R-project.org/package=SimilaR
files <- list.files(system.file("examples", package="rscc"), "*.R$", full.names=TRUE)
prgs  <- sourcecode(files)
simil <- similarities(prgs)
if (interactive()) browse(prgs, simil)
}
